StemSepApp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.37.1
antlr4-python3-runtime==4.9.3
anyio==4.11.0
Appium-Python-Client==5.2.4
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
audioread==3.0.1
auto-py-to-exe==2.43.0
blessed==1.24.0
bottle==0.13.4

# Required by StemSepApp/src/audio/simple_separator.py (audio-separator integration).
# This is the Python package from https://github.com/nomadkaraoke/python-audio-separator
audio-separator

# CuPy is optional (GPU acceleration) and is often unavailable for newer Python versions (e.g. 3.14) on Windows.
# Install manually if needed, pinned to a compatible CUDA/Python combo.
cupy-cuda12x==13.6.0; python_version < "3.14"
customtkinter==5.2.2
cycler==0.12.1
Cython==3.2.1
darkdetect==0.8.0
debugpy==1.8.17
decorator==5.2.1
demucs==4.0.0
diffq
distro==1.9.0
docstring_parser==0.17.0
docutils==0.22.3
dora_search==0.1.12
editor==1.6.6
Eel==0.18.2
einops==0.8.1
executing==2.2.1
fastrlock==0.8.3
fastuuid==0.14.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.50.1
google-generativeai==0.7.2
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
GPUtil==1.4.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.20.0
humanfriendly==10.0
id==1.5.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
inquirer==3.4.1
ipykernel==6.31.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==25.7.0
kiwisolver==1.4.9
# lameenc is optional (MP3 encoding) and may not have wheels for newer Python versions (e.g. 3.14) on Windows.
# FFmpeg-based MP3 export is supported elsewhere in the stack; install lameenc manually if you need it and it is compatible.
lameenc==1.8.1; python_version < "3.14"
lazy_loader==0.4
librosa==0.11.0
litellm==1.80.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.2
nltk==3.9.2
# numba is optional but can significantly speed up some DSP paths.
# On Python 3.14+ (especially on Windows) numba may refuse to install due to version constraints.
# Librosa can run without numba (potentially slower). Install numba manually on a supported Python.
numba==0.62.1; python_version < "3.14"
numpy
omegaconf==2.3.0
onnx==1.19.1
# onnxruntime is optional (ONNX inference) and may not have wheels for newer Python versions (e.g. 3.14) on Windows.
# CPU-only non-ONNX flows can run without it. Install onnxruntime manually on a supported Python if you need ONNX models.
onnxruntime==1.23.2; python_version < "3.14"
open-interpreter==0.4.3
openai
openunmix==1.3.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pefile==2024.8.26
Pillow
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
progress==1.6.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.23
pydantic
pydantic_core==2.41.5
pydub==0.25.1
pygame
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller
pyinstaller-hooks-contrib==2025.9
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==27.1.0
readchar==4.2.1
readme_renderer==44.0
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.4.2
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
runs==1.2.2
scikit-learn
scipy
selenium
Send2Trash==1.8.3
setuptools
shapely==2.1.2
shellingham==1.5.4
shortuuid==1.0.13
simpleaudio

six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.37.2
static_ffmpeg==2.13
submitit==1.5.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken

tkinter-tooltip==2.1.0
tokenizers==0.22.1
tokentrim==0.1.13
toml==0.10.2

tornado==6.5.2
tqdm==4.66.1
traitlets==5.14.3
treetable==0.2.6
trio==0.32.0
trio-websocket==0.12.2
twine==6.2.0
typer==0.12.5
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wget==3.2
wheel==0.45.1
wsproto==1.3.1
xmod==1.8.1
yarl==1.22.0
yaspin==3.3.0
yt-dlp
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1

mutagen

beartype
rotary-embedding-torch
py-cpuinfo
